requirements.txt
selenium==4.36.0
pytest==8.4.2
allure-pytest==2.15.3